Modulo2 (#19)
* Modify data

* Modify data

* Modify data

* Modify Controllers

Modify Controllers

* Mod Routes

Mod Routes

* Modify data

* Modify data

* Modify data

* Modify data

* ControllersZ

* ControllersZ

* ControllersZ

* Modify data

* ControllersZ

* Modify data

* Nuevas tablas

* ControllersZ

* Modify data

* Modify data

* ControllersZ

* CambiosZ

* CambiosZ

* CambiosZ

* CambiosZ

* Modify data

+++ database/migrations/2023_11_02_154701_create_tb_riesgo_arl_table.php
@@ -16,7 +16,7 @@ class CreateTbRiesgoArlTable extends Migration
         Schema::create('tb_riesgo_arl', function (Blueprint $table) {
             $table->id();
             $table->string('riesgo');
-            $table->float('porcentaje');
+            $table->double('porcentaje');
             $table->boolean('estado')->default(1);
         });
     }
